requirements.txt
flask==2.3.3
transformers==4.41.1
torch==2.3.0
Pillow==10.3.0
gTTS==2.5.1
requests==2.31.0


